pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["NIK <nikita1z@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.0.1"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
django-filter = "^23.1"
django-environ = "^0.10.0"
django-cors-headers = "^3.14.0"
whitenoise = "^6.4.0"
optional-django = "^0.3.0"
django-taggit = "^3.1.0"
markdown = "3.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
